+++ src/mongoChangeStream.test.ts
@@ -7,7 +7,6 @@ import assert from 'node:assert'
 import { initSync } from './mongoChangeStream.js'
 import {
   JSONSchema,
-  SyncOptions,
   SchemaChangeEvent,
   ScanOptions,
   ChangeStreamOptions,
@@ -24,16 +23,22 @@ import ms from 'ms'
 import { setTimeout } from 'node:timers/promises'
 import { QueueOptions } from 'prom-utils'
 
-const init = _.memoize(async (options?: SyncOptions) => {
+const getConns = _.memoize(async (x?: any) => {
+  console.log(x)
   const redis = new Redis({ keyPrefix: 'testing:' })
   const client = await MongoClient.connect(process.env.MONGO_CONN as string)
   const db = client.db()
   const coll = db.collection('testing')
-  const sync = initSync(redis, coll, options)
-  sync.emitter.on('stateChange', console.log)
 
-  return { sync, db, coll, redis }
+  return { client, db, coll, redis }
 })
+
+const getSync = async () => {
+  const { redis, coll } = await getConns()
+  const sync = initSync(redis, coll)
+  sync.emitter.on('stateChange', console.log)
+  return sync
+}
 
 const genUser = () => ({
   name: faker.name.fullName(),
@@ -59,16 +64,18 @@ const schema: JSONSchema = {
 
 const numDocs = 500
 
-const populateCollection = (collection: Collection) => {
+const populateCollection = (collection: Collection, count = numDocs) => {
   const users = []
-  for (let i = 0; i < numDocs; i++) {
+  for (let i = 0; i < count; i++) {
     users.push({ insertOne: { document: genUser() } })
   }
   return collection.bulkWrite(users)
 }
 
-const before = async () => {
-  const { sync, coll } = await init()
+const initState = async (
+  sync: ReturnType<typeof initSync>,
+  coll: Collection
+) => {
   // Reset state
   await sync.reset()
   await coll.deleteMany({})
@@ -77,8 +84,9 @@ const before = async () => {
 }
 
 test('should complete initial scan', async () => {
-  const { sync } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   const processed = []
   const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
@@ -94,33 +102,99 @@ test('should complete initial scan', async () => {
   await initialScan.stop()
 })
 
-test('initial scan should resume properly', async () => {
-  const { sync } = await init()
-  await before()
+test('should complete initial scan if collection is empty', async () => {
+  const { coll } = await getConns()
+  const sync = await getSync()
+
+  // Reset state
+  await sync.reset()
+  await coll.deleteMany({})
 
   const processed = []
   const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
     await setTimeout(50)
     processed.push(...docs)
   }
+  let completed = false
+  sync.emitter.on('initialScanComplete', () => {
+    completed = true
+  })
   const scanOptions = { batchSize: 100 }
+  const initialScan = await sync.runInitialScan(processRecords, scanOptions)
+  // Wait for initial scan to complete
+  await initialScan.start()
+  assert.ok(completed)
+  assert.equal(processed.length, 0)
+  // Stop
+  await initialScan.stop()
+})
+
+test('initial scan should resume after stop', async () => {
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
+
+  const processed = []
+  const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
+    await setTimeout(10)
+    processed.push(...docs)
+  }
+  const scanOptions = { batchSize: 50 }
+  const initialScan = await sync.runInitialScan(processRecords, scanOptions)
+  let completed = false
+  sync.emitter.on('initialScanComplete', () => {
+    completed = true
+  })
+  // Start
+  initialScan.start()
+  // Allow for some records to be processed
+  await setTimeout(500)
+  // Stop the initial scan
+  await initialScan.stop()
+  // Wait for the initial scan to complete
+  initialScan.start()
+  // Add some more records
+  await populateCollection(coll, 10)
+  await setTimeout(ms('5s'))
+  assert.ok(completed)
+  assert.equal(processed.length, numDocs + 10)
+  // Stop
+  await initialScan.stop()
+})
+
+test('initial scan should not be marked as completed if connection is closed', async () => {
+  // Memoize hack
+  const { coll, redis, client } = await getConns({})
+  const sync = initSync(redis, coll)
+  sync.emitter.on('stateChange', console.log)
+  await initState(sync, coll)
+
+  const processed = []
+  const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
+    await setTimeout(10)
+    processed.push(...docs)
+  }
+  const scanOptions = { batchSize: 50 }
   const initialScan = await sync.runInitialScan(processRecords, scanOptions)
   // Start
   initialScan.start()
   // Allow for some records to be processed
-  await setTimeout(50)
-  // Stop the initial scan
-  await initialScan.stop()
-  // Wait for the initial scan to complete
-  await initialScan.start()
-  assert.equal(processed.length, numDocs)
+  await setTimeout(200)
+  // Close the connection.
+  await client.close()
+  // Allow for some time
+  await setTimeout(100)
+  // Check if completed
+  const completedAt = await redis.get(sync.keys.scanCompletedKey)
+  assert.equal(completedAt, null)
   // Stop
   await initialScan.stop()
 })
 
 test('should process records via change stream', async () => {
-  const { sync, coll } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   const processed = []
   const processRecord = async (doc: ChangeStreamDocument) => {
@@ -141,8 +215,9 @@ test('should process records via change stream', async () => {
 })
 
 test('change stream should resume properly', async () => {
-  const { sync, coll } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   const processed = []
   // Change stream
@@ -170,7 +245,7 @@ test('change stream should resume properly', async () => {
 })
 
 test('starting change stream is idempotent', async () => {
-  const { sync } = await init()
+  const sync = await getSync()
   // Change stream
   const processRecord = async () => {
     await setTimeout(5)
@@ -183,8 +258,9 @@ test('starting change stream is idempotent', async () => {
 })
 
 test('stopping change stream is idempotent', async () => {
-  const { sync, coll } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   // Change stream
   const processRecord = async () => {
@@ -200,8 +276,9 @@ test('stopping change stream is idempotent', async () => {
 })
 
 test('starting initial scan is idempotent', async () => {
-  const { sync } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   const processRecords = async () => {
     await setTimeout(50)
@@ -214,8 +291,9 @@ test('starting initial scan is idempotent', async () => {
 })
 
 test('stopping initial scan is idempotent', async () => {
-  const { sync } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   const processRecords = async () => {
     await setTimeout(50)
@@ -229,8 +307,9 @@ test('stopping initial scan is idempotent', async () => {
 })
 
 test('Should resync when resync flag is set', async () => {
-  const { sync, redis } = await init()
-  await before()
+  const { coll, redis } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   let resyncTriggered = false
   const processed = []
@@ -272,7 +351,7 @@ test('Should resync when resync flag is set', async () => {
 })
 
 test('Resync start/stop is idempotent', async () => {
-  const { sync } = await init()
+  const sync = await getSync()
 
   const resync = sync.detectResync()
   resync.start()
@@ -282,7 +361,8 @@ test('Resync start/stop is idempotent', async () => {
 })
 
 test('Detect schema change', async () => {
-  const { db, coll, sync } = await init()
+  const { db, coll } = await getConns()
+  const sync = await getSync()
   // Set schema
   await db.command({
     collMod: coll.collectionName,
@@ -311,7 +391,8 @@ test('Detect schema change', async () => {
 })
 
 test('Schema change start/stop is idempotent', async () => {
-  const { sync, db } = await init()
+  const { db } = await getConns()
+  const sync = await getSync()
 
   const schemaChange = await sync.detectSchemaChange(db)
   schemaChange.start()
@@ -321,8 +402,9 @@ test('Schema change start/stop is idempotent', async () => {
 })
 
 test('should fail health check - initial scan', async () => {
-  const { sync } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   let healthCheckFailed = false
   const processed = []
@@ -349,8 +431,9 @@ test('should fail health check - initial scan', async () => {
 })
 
 test('should fail health check - change stream', async () => {
-  const { sync, coll } = await init()
-  await before()
+  const { coll } = await getConns()
+  const sync = await getSync()
+  await initState(sync, coll)
 
   let healthCheckFailed = false
   const processed = []